+++ migrations/m200229_131724_blog_posts.php
@@ -3,9 +3,9 @@
 use yii\db\Migration;
 
 /**
- * Class m200229_131724_blog
+ * Class m200229_131724_blog_posts
  */
-class m200229_131724_blog extends Migration
+class m200229_131724_blog_posts extends Migration
 {
     /**
      * {@inheritdoc}
@@ -18,7 +18,7 @@ class m200229_131724_blog extends Migration
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
         }
 
-        $this->createTable('{{%blog}}', [
+        $this->createTable('{{%blog_posts}}', [
 
             'id' => $this->bigPrimaryKey(),
             'name' => $this->string(128)->notNull(),
@@ -50,18 +50,18 @@ class m200229_131724_blog extends Migration
 
         ], $tableOptions);
 
-        $this->createIndex('{{%idx-blog-alias}}', '{{%blog}}', ['name', 'alias']);
-        $this->createIndex('{{%idx-blog-status}}', '{{%blog}}', ['alias', 'status']);
-        $this->createIndex('{{%idx-blog-content}}','{{%blog}}', ['name', 'excerpt', 'content(250)'],false);
+        $this->createIndex('{{%idx-blog-posts-alias}}', '{{%blog_posts}}', ['name', 'alias']);
+        $this->createIndex('{{%idx-blog-posts-status}}', '{{%blog_posts}}', ['alias', 'status']);
+        $this->createIndex('{{%idx-blog-posts-content}}','{{%blog_posts}}', ['name', 'excerpt', 'content(250)'],false);
 
         // If exist module `Users` set foreign key `created_by`, `updated_by` to `users.id`
         if (class_exists('\wdmg\users\models\Users')) {
-            $this->createIndex('{{%idx-blog-created}}','{{%blog}}', ['created_by'],false);
-            $this->createIndex('{{%idx-blog-updated}}','{{%blog}}', ['updated_by'],false);
+            $this->createIndex('{{%idx-blog-posts-created}}','{{%blog_posts}}', ['created_by'],false);
+            $this->createIndex('{{%idx-blog-posts-updated}}','{{%blog_posts}}', ['updated_by'],false);
             $userTable = \wdmg\users\models\Users::tableName();
             $this->addForeignKey(
-                'fk_blog_to_users1',
-                '{{%blog}}',
+                'fk_blog_posts_to_users1',
+                '{{%blog_posts}}',
                 'created_by',
                 $userTable,
                 'id',
@@ -69,8 +69,8 @@ class m200229_131724_blog extends Migration
                 'CASCADE'
             );
             $this->addForeignKey(
-                'fk_blog_to_users2',
-                '{{%blog}}',
+                'fk_blog_posts_to_users2',
+                '{{%blog_posts}}',
                 'updated_by',
                 $userTable,
                 'id',
@@ -85,28 +85,28 @@ class m200229_131724_blog extends Migration
      */
     public function safeDown()
     {
-        $this->dropIndex('{{%idx-blog-alias}}', '{{%blog}}');
-        $this->dropIndex('{{%idx-blog-status}}', '{{%blog}}');
-        $this->dropIndex('{{%idx-blog-content}}', '{{%blog}}');
+        $this->dropIndex('{{%idx-blog-posts-alias}}', '{{%blog_posts}}');
+        $this->dropIndex('{{%idx-blog-posts-status}}', '{{%blog_posts}}');
+        $this->dropIndex('{{%idx-blog-posts-content}}', '{{%blog_posts}}');
 
         if(class_exists('\wdmg\users\models\Users')) {
-            $this->dropIndex('{{%idx-blog-created}}', '{{%blog}}');
-            $this->dropIndex('{{%idx-blog-updated}}', '{{%blog}}');
+            $this->dropIndex('{{%idx-blog-posts-created}}', '{{%blog_posts}}');
+            $this->dropIndex('{{%idx-blog-posts-updated}}', '{{%blog_posts}}');
             $userTable = \wdmg\users\models\Users::tableName();
             if (!(Yii::$app->db->getTableSchema($userTable, true) === null)) {
                 $this->dropForeignKey(
-                    'fk_blog_to_users1',
-                    '{{%blog}}'
+                    'fk_blog_posts_to_users1',
+                    '{{%blog_posts}}'
                 );
                 $this->dropForeignKey(
-                    'fk_blog_to_users2',
-                    '{{%blog}}'
+                    'fk_blog_posts_to_users2',
+                    '{{%blog_posts}}'
                 );
             }
         }
 
-        $this->truncateTable('{{%blog}}');
-        $this->dropTable('{{%blog}}');
+        $this->truncateTable('{{%blog_posts}}');
+        $this->dropTable('{{%blog_posts}}');
     }
 
 }