+++ views/cats/index.php
@@ -34,6 +34,7 @@ $this->params['breadcrumbs'][] = $this->title;
                 'format' => 'raw',
                 'value' => function($model) {
                     $output = Html::tag('strong', $model->name);
+                    $output .= (($model->id === 1) ? " <span class=\"text-muted\">(" . Yii::t('app/modules/blog', 'default') . ")</span>" : "");
                     if (($categoryURL = $model->getCategoryUrl(true, true)) && $model->id) {
                         $output .= '<br/>' . Html::a($model->getCategoryUrl(true, false), $categoryURL, [
                                 'target' => '_blank',
@@ -117,6 +118,11 @@ $this->params['breadcrumbs'][] = $this->title;
                 'contentOptions' => [
                     'class' => 'text-center'
                 ],
+                'visibleButtons' => [
+                    'delete' => function ($data) {
+                        return !($data->id === 1); // Category for uncategorized posts (undeleted).
+                    },
+                ]
             ]
         ]
     ]); ?>