+++ views/default/index.php
@@ -6,7 +6,7 @@ use yii\helpers\Url;
 use yii\widgets\ListView;
 
 /* @var $this yii\web\View */
-/* @var $model wdmg\blog\models\Blog */
+/* @var $model wdmg\blog\models\Posts */
 
 echo ListView::widget([
     'dataProvider' => $dataProvider,